requirements.txt
pandas
yfinance
ta
line-bot-sdk
python-dotenv
schedule
